\name{mzR-package}
\alias{mzR-package}
\alias{mzR}
\docType{package}
\title{
parser for mzXML, mzData and mzML files (mass spectrometry data)
}
\description{
R frontend for the ISB random acces parser for mass
spectrometry mzXML, mzData and mzML files. The package
contains the original code written by the ISB, and a
subset of the proteowizard library for mzML.
}
\details{
\tabular{ll}{
Package: \tab mzR\cr
Type: \tab Package\cr
Version: \tab 0.4.3\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
Depends: \tab methods, Rcpp (>= 0.9.4)\cr
Imports: \tab Biobase\cr
Suggests: \tab msdata, RUnit, faahKO\cr
LinkingTo: \tab Rcpp\cr
RcppModules: \tab Ramp\cr
SystemRequirements: \tab GNU make, NetCDF, zlib\cr
biocViews: \tab Infrastructure, Bioinformatics, DataImport, Proteomics, MassSpectrometry\cr
Built: \tab R 2.13.0; x86_64-pc-linux-gnu; 2011-05-28 19:11:38 UTC; unix\cr
}

Index:
\preformatted{
openMSfile              No documentation yet
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{mzR} \tab mzR, Ramp, mzXML, mzData, mzML (source, pdf)\cr
}
}
\author{
Bernd Fischer, Steffen Neumann, Laurent Gatto

Maintainer: Bernd Fischer <bernd.fischer@embl.de>,
Steffen Neumann <sneumann@ipb-halle.de>,
Laurent Gatto <lg390@cam.ac.uk>
}
\references{
  http://dx.doi.org/10.1074/mcp.R110.000133
}
\keyword{package, file}
