\name{mzXML-package}
\alias{mzXML-package}
\alias{mzXML}
\docType{package}
\title{
A parser for mass spectrometry data in the format mzXML.
}
\description{
The package contains a class for that handles mzXML file parsing.
}
\details{
The package contains the class \code{\link{Ramp}}. Please refer to the class documentation.

\tabular{ll}{
Package: \tab mzXML\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-17\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}

}
\author{
Bernd Fischer

Maintainer:Bernd Fischer <bernd.fischer@embl.de>
}
\references{
}
\keyword{ package, mzXML, mass spectrometry, proteomics }
\seealso{
\code{\link{Ramp}}
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
