\name{mzR-package}
\alias{mzR-package}
\alias{mzR}
\docType{package}
\title{
A parser for mass spectrometry data in the formats mzXML, mzData and mzML.
}
\description{
The package contains a class for that handles mass spectrometry XML file parsing.
}
\details{
The package contains the class \code{\link{Ramp}}. Please refer to the class documentation.

\tabular{ll}{
Package: \tab mzR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-17\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}

}
\author{
Bernd Fischer, Steffen Neumann

Maintainer:Bernd Fischer <bernd.fischer@embl.de>, Steffen Neumann <sneumann@ipb-halle.de>
}
\references{
}
\keyword{package, mzXML, mzData, mzML, mass spectrometry, proteomics, metabolomics }
\seealso{
\code{\link{Ramp}}
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
